+++ internal/file/proto_set_provider_test.go
@@ -84,7 +84,6 @@ func TestProtoSetProviderGetMultipleForFilesAll(t *testing.T) {
 					ExcludePrefixes: []string{
 						cwd + "/testdata/c/i",
 						cwd + "/testdata/d",
-						cwd + "/testdata/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.4.0",
@@ -126,7 +125,6 @@ func TestProtoSetProviderGetMultipleForFilesAll(t *testing.T) {
 					DirPath: cwd + "/testdata/a/d",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/a/d/file3.proto",
-						cwd + "/testdata/a/d/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.2.0",
@@ -166,7 +164,6 @@ func TestProtoSetProviderGetMultipleForFilesAll(t *testing.T) {
 					DirPath: cwd + "/testdata/b",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/b/g/h",
-						cwd + "/testdata/b/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.3.0",
@@ -242,7 +239,6 @@ func TestProtoSetProviderGetMultipleForFilesSomeMissing(t *testing.T) {
 					ExcludePrefixes: []string{
 						cwd + "/testdata/c/i",
 						cwd + "/testdata/d",
-						cwd + "/testdata/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.4.0",
@@ -280,7 +276,6 @@ func TestProtoSetProviderGetMultipleForFilesSomeMissing(t *testing.T) {
 					DirPath: cwd + "/testdata/a/d",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/a/d/file3.proto",
-						cwd + "/testdata/a/d/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.2.0",
@@ -348,7 +343,6 @@ func TestProtoSetProviderGetMultipleForDirCwdRel(t *testing.T) {
 					ExcludePrefixes: []string{
 						cwd + "/testdata/c/i",
 						cwd + "/testdata/d",
-						cwd + "/testdata/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.4.0",
@@ -386,7 +380,6 @@ func TestProtoSetProviderGetMultipleForDirCwdRel(t *testing.T) {
 					DirPath: cwd + "/testdata/a/d",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/a/d/file3.proto",
-						cwd + "/testdata/a/d/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.2.0",
@@ -420,7 +413,6 @@ func TestProtoSetProviderGetMultipleForDirCwdRel(t *testing.T) {
 					DirPath: cwd + "/testdata/b",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/b/g/h",
-						cwd + "/testdata/b/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.3.0",
@@ -488,7 +480,6 @@ func TestProtoSetProviderGetMultipleForDirCwdAbs(t *testing.T) {
 					ExcludePrefixes: []string{
 						cwd + "/testdata/c/i",
 						cwd + "/testdata/d",
-						cwd + "/testdata/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.4.0",
@@ -526,7 +517,6 @@ func TestProtoSetProviderGetMultipleForDirCwdAbs(t *testing.T) {
 					DirPath: cwd + "/testdata/a/d",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/a/d/file3.proto",
-						cwd + "/testdata/a/d/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.2.0",
@@ -560,7 +550,6 @@ func TestProtoSetProviderGetMultipleForDirCwdAbs(t *testing.T) {
 					DirPath: cwd + "/testdata/b",
 					ExcludePrefixes: []string{
 						cwd + "/testdata/b/g/h",
-						cwd + "/testdata/b/vendor",
 					},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.3.0",
@@ -611,10 +600,8 @@ func TestProtoSetProviderGetMultipleForDirCwdSubRel(t *testing.T) {
 					},
 				},
 				Config: settings.Config{
-					DirPath: cwd + "/testdata/d",
-					ExcludePrefixes: []string{
-						cwd + "/testdata/d/vendor",
-					},
+					DirPath:         cwd + "/testdata/d",
+					ExcludePrefixes: []string{},
 					Compile: settings.CompileConfig{
 						ProtobufVersion:       "3.3.0",
 						IncludePaths:          []string{},